refactor(Req): add requirements analysis workview page
