--- src/ramstk/gui/gtk/workviews/Requirement.py
@@ -1,1253 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-#       ramstk.gui.gtk.workviews.Requirement.py is part of the RAMSTK Project
-#
-# All rights reserved.
-# Copyright 2007 - 2017 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
-"""The RAMSTK Requirement Work View."""
-
-# Third Party Imports
-from pubsub import pub
-
-# RAMSTK Package Imports
-from ramstk import boolean_to_integer
-from ramstk.gui.gtk.ramstk import (
-    RAMSTKButton, RAMSTKCheckButton, RAMSTKComboBox,
-    RAMSTKDateSelect, RAMSTKEntry, RAMSTKFrame,
-    RAMSTKLabel, RAMSTKScrolledWindow, do_make_label_group
-)
-from ramstk.gui.gtk.ramstk.Widget import Gdk, GObject, Gtk, Pango, _
-
-# RAMSTK Local Imports
-from .WorkView import RAMSTKWorkView
-
-
-class GeneralData(RAMSTKWorkView):
-    """
-    Display Requirement attribute data in the RAMSTK Work Book.
-
-    The Work View displays all the attributes for the selected Requirement. The
-    attributes of a Work View are:
-
-    :ivar int _requirement_id: the ID of the Requirement Data Model currently
-                               being displayed.
-    :ivar chkDerived: the :class:`Gtk.CheckButton` used to indicate the
-                      selected Requirement is derived.
-    :ivar chkValidated: the :class:`Gtk.CheckButton` used to indicates the
-                        selected Requirement has been validated.
-    :ivar cmbOwner: the :class:`Gtk.ComboBox` used to display/select the
-                    owning organization for the Requirement.
-    :ivar cmbRequirementType: the :class:`Gtk.ComboBox` used to
-                              display/select the type of Requirement.
-    :ivar cmbPriority: the :class:`Gtk.ComboBox` used to display/select the
-                       priority of the Requirement.
-    :ivar txtFigNum: the :class:`Gtk.Entry` used to display/enter the
-                     spcification figure number associated with the
-                     Requirement.
-    :ivar txtPageNum: the :class:`Gtk.Entry` used to display/enter the
-                      specification page number associated with the
-                      Requirement.
-    :ivar txtSpecification: the :class:`Gtk.Entry` used to display/enter the
-                            governing specification.
-    :ivar txtValidatedDate: the :class:`Gtk.Entry` used to display/enter the
-                            Requirement was validated.
-
-    Callbacks signals in _lst_handler_id:
-
-    +----------+-------------------------------------------+
-    | Position | Widget - Signal                           |
-    +==========+===========================================+
-    |     0    | txtCode `changed`                         |
-    +----------+-------------------------------------------+
-    |     1    | txtName `changed`                         |
-    +----------+-------------------------------------------+
-    |     2    | cmbRequirementType `changed`              |
-    +----------+-------------------------------------------+
-    |     3    | chkDerived `toggled`                      |
-    +----------+-------------------------------------------+
-    |     4    | txtSpcification `changed`                 |
-    +----------+-------------------------------------------+
-    |     5    | txtPageNum `changed`                      |
-    +----------+-------------------------------------------+
-    |     6    | txtFigNum `changed`                       |
-    +----------+-------------------------------------------+
-    |     7    | cmbPriority `changed`                     |
-    +----------+-------------------------------------------+
-    |     8    | cmbOwner `changed`                        |
-    +----------+-------------------------------------------+
-    |     9    | chKValidated `toggled`                    |
-    +----------+-------------------------------------------+
-    |    10    | txtValidateDate `changed`                 |
-    +----------+-------------------------------------------+
-    """
-
-    _lst_labels = [
-        _("Requirement Code:"),
-        _("Requirement Description:"),
-        _("Requirement Type:"), "",
-        _("Specification:"),
-        _("Page Number:"),
-        _("Figure Number:"),
-        _("Priority:"),
-        _("Owner:"), "",
-        _("Validated Date:"),
-    ]
-
-    def __init__(self, configuration, **kwargs):  # pylint: disable=unused-argument
-        """
-        Initialize the Requirement Work View.
-
-        :param configuration: the RAMSTK Configuration class instance.
-        :type configuration: :class:`ramstk.Configuration.Configuration`
-        """
-        RAMSTKWorkView.__init__(self, configuration, module='Requirement')
-
-        # Initialize private dictionary attributes.
-
-        # Initialize private list attributes.
-
-        # Initialize private scalar attributes.
-        self._requirement_id = None
-
-        # Initialize public dictionary attributes.
-
-        # Initialize public list attributes.
-
-        # Initialize public scalar attributes.
-        self.btnValidateDate = RAMSTKButton(label="...")
-
-        self.chkDerived = RAMSTKCheckButton(label=_("Requirement is derived."))
-        self.chkValidated = RAMSTKCheckButton(
-            label=_("Requirement is validated."),
-        )
-
-        self.cmbOwner = RAMSTKComboBox()
-        self.cmbRequirementType = RAMSTKComboBox(index=1, simple=False)
-        self.cmbPriority = RAMSTKComboBox()
-
-        self.txtCode = RAMSTKEntry()
-        self.txtFigNum = RAMSTKEntry()
-        self.txtName = RAMSTKEntry()
-        self.txtPageNum = RAMSTKEntry()
-        self.txtSpecification = RAMSTKEntry()
-        self.txtValidatedDate = RAMSTKEntry()
-
-        self.__set_properties()
-        self.__load_combobox()
-        self.__make_ui()
-        self.__set_callbacks()
-
-        # Subscribe to PyPubSub messages.
-        pub.subscribe(self._do_clear_page, 'closed_program')
-        pub.subscribe(self._do_load_page, 'selected_requirement')
-        pub.subscribe(self._do_load_code, 'created_requirement_code')
-        pub.subscribe(self._on_edit, 'mvw_editing_requirement')
-
-    def __load_combobox(self):
-        """
-        Load the RAMSTK ComboBox widgets with lists of information.
-
-        :return: None
-        :rtype: None
-        """
-        # Load the requirement type Gtk.ComboBox(); each _type is
-        # (Code, Description, Type).
-        _types = []
-        for _index, _key in enumerate(
-                self.RAMSTK_CONFIGURATION.RAMSTK_REQUIREMENT_TYPE,
-        ):
-            _types.append(
-                self.RAMSTK_CONFIGURATION.RAMSTK_REQUIREMENT_TYPE[_key],
-            )
-        self.cmbRequirementType.do_load_combo(
-            list(_types),
-            index=1,
-            simple=False,
-        )
-
-        # Load the owner Gtk.ComboBox(); each _owner is
-        # (Description, Group Type).
-        _owners = []
-        for _index, _key in enumerate(
-                self.RAMSTK_CONFIGURATION.RAMSTK_WORKGROUPS,
-        ):
-            _owners.append(self.RAMSTK_CONFIGURATION.RAMSTK_WORKGROUPS[_key])
-        self.cmbOwner.do_load_combo(list(_owners))
-
-        # Load the priority Gtk.Combo().
-        _priorities = [["1"], ["2"], ["3"], ["4"], ["5"]]
-        self.cmbPriority.do_load_combo(_priorities)
-
-    def __make_ui(self):
-        """
-        Make the Requirement Work View General Data Gtk.Notebook() page.
-
-        :return: None
-        :rtype: None
-        """
-        _scrolledwindow = Gtk.ScrolledWindow()
-        _scrolledwindow.set_policy(
-            Gtk.PolicyType.NEVER,
-            Gtk.PolicyType.AUTOMATIC,
-        )
-        _scrolledwindow.add_with_viewport(
-            RAMSTKWorkView._make_buttonbox(
-                self,
-                icons=[],
-                tooltips=[],
-                callbacks=[],
-            ),
-        )
-        self.pack_start(_scrolledwindow, False, False, 0)
-
-        _fixed = Gtk.Fixed()
-
-        _scrollwindow = RAMSTKScrolledWindow(_fixed)
-        _frame = RAMSTKFrame(label=_("General Information"))
-        _frame.add(_scrollwindow)
-
-        _x_pos, _y_pos = do_make_label_group(self._lst_labels, _fixed, 5, 5)
-        _x_pos += 50
-
-        _fixed.put(self.txtCode, _x_pos, _y_pos[0])
-        _fixed.put(self.txtName, _x_pos, _y_pos[1])
-        _fixed.put(self.cmbRequirementType, _x_pos, _y_pos[2])
-        _fixed.put(self.chkDerived, _x_pos, _y_pos[3] + 5)
-        _fixed.put(self.txtSpecification, _x_pos, _y_pos[4])
-        _fixed.put(self.txtPageNum, _x_pos, _y_pos[5])
-        _fixed.put(self.txtFigNum, _x_pos, _y_pos[6])
-        _fixed.put(self.cmbPriority, _x_pos, _y_pos[7])
-        _fixed.put(self.cmbOwner, _x_pos, _y_pos[8])
-        _fixed.put(self.chkValidated, _x_pos, _y_pos[9] + 5)
-        _fixed.put(self.txtValidatedDate, _x_pos, _y_pos[10])
-        _fixed.put(self.btnValidateDate, _x_pos + 205, _y_pos[10])
-
-        self.pack_start(_frame, True, True, 0)
-
-        # Create the label for the Gtk.Notebook() tab.
-        _label = RAMSTKLabel(
-            _("General\nData"),
-            height=30,
-            width=-1,
-            justify=Gtk.Justification.CENTER,
-            tooltip=_(
-                "Displays general information for the selected "
-                "requirement.",
-            ),
-        )
-        self.hbx_tab_label.pack_start(_label, True, True, 0)
-
-        self.show_all()
-
-    def __set_callbacks(self):
-        """
-        Set the callback methods and functions.
-
-        :return: None
-        :rtype: None
-        """
-        self._lst_handler_id.append(
-            self.txtCode.connect('changed', self._on_focus_out, 0),
-        )
-        self._lst_handler_id.append(
-            self.txtName.connect('changed', self._on_focus_out, 1),
-        )
-        self._lst_handler_id.append(
-            self.cmbRequirementType.connect(
-                'changed', self._on_combo_changed,
-                2,
-            ),
-        )
-        self._lst_handler_id.append(
-            self.chkDerived.connect('toggled', self._on_toggled, 3),
-        )
-        self._lst_handler_id.append(
-            self.txtSpecification.connect('changed', self._on_focus_out, 4),
-        )
-        self._lst_handler_id.append(
-            self.txtPageNum.connect('changed', self._on_focus_out, 5),
-        )
-        self._lst_handler_id.append(
-            self.txtFigNum.connect('changed', self._on_focus_out, 6),
-        )
-        self._lst_handler_id.append(
-            self.cmbPriority.connect('changed', self._on_combo_changed, 7),
-        )
-        self._lst_handler_id.append(
-            self.cmbOwner.connect('changed', self._on_combo_changed, 8),
-        )
-        self._lst_handler_id.append(
-            self.chkValidated.connect('toggled', self._on_toggled, 9),
-        )
-        self._lst_handler_id.append(
-            self.txtValidatedDate.connect('changed', self._on_focus_out, 10),
-        )
-        self._lst_handler_id.append(
-            self.btnValidateDate.connect(
-                'button-release-event',
-                self._do_select_date,
-                self.txtValidatedDate,
-            ),
-        )
-
-    def __set_properties(self):
-        """
-        Set the properties of the General Data Work View and widgets.
-
-        :return: None
-        :rtype: None
-        """
-        # ----- BUTTONS
-        self.btnValidateDate.do_set_properties(height=25, width=25)
-
-        self.chkDerived.do_set_properties(
-            tooltip=_(
-                "Indicates whether or not the selected requirement is "
-                "derived.",
-            ),
-        )
-        self.chkValidated.do_set_properties(
-            tooltip=_(
-                "Indicates whether or not the selected requirement is "
-                "validated.",
-            ),
-        )
-
-        # ----- COMBOBOXES
-        self.cmbPriority.do_set_properties(width=50)
-
-        # ----- ENTRIES
-        self.txtCode.do_set_properties(
-            width=125,
-            tooltip=_("A unique code for the selected requirement."),
-        )
-        self.txtName.do_set_properties(
-            height=100,
-            width=800,
-            tooltip=_("The description of the selected requirement."),
-        )
-
-    def _do_clear_page(self):
-        """
-        Clear the contents of the page.
-
-        :return: None
-        :rtype: None
-        """
-        self.txtCode.do_update('', self._lst_handler_id[0])
-        self.txtName.do_update('', self._lst_handler_id[1])
-
-        self.cmbRequirementType.handler_block(self._lst_handler_id[2])
-        self.cmbRequirementType.set_active(0)
-        self.cmbRequirementType.handler_unblock(self._lst_handler_id[2])
-
-        self.chkDerived.do_update(False, self._lst_handler_id[3])
-        self.txtSpecification.do_update('', self._lst_handler_id[4])
-        self.txtPageNum.do_update('', self._lst_handler_id[5])
-        self.txtFigNum.do_update('', self._lst_handler_id[6])
-
-        self.cmbPriority.handler_block(self._lst_handler_id[7])
-        self.cmbPriority.set_active(0)
-        self.cmbPriority.handler_unblock(self._lst_handler_id[7])
-
-        self.cmbOwner.handler_block(self._lst_handler_id[8])
-        self.cmbOwner.set_active(0)
-        self.cmbOwner.handler_unblock(self._lst_handler_id[8])
-
-        self.chkValidated.do_update(False, self._lst_handler_id[9])
-        self.txtValidatedDate.do_update('', self._lst_handler_id[10])
-
-    def _do_load_code(self, code):
-        """
-        Load the Requirement code RAMSTKEntry().
-
-        :param str code: the Requirement code to load.
-        :return: None
-        :rtype: None
-        """
-        self.txtCode.do_update(str(code), self._lst_handler_id[0])
-
-    def _do_load_page(self, attributes):
-        """
-        Load the Requirements general data page.
-
-        :return: None
-        :rtype: None
-        """
-        self._revision_id = attributes['revision_id']
-        self._requirement_id = attributes['requirement_id']
-
-        RAMSTKWorkView.on_select(
-            self,
-            title=_("Analyzing Requirement {0:s} - {1:s}").format(
-                str(attributes['requirement_code']),
-                str(attributes['description']),
-            ),
-        )
-
-        self.txtCode.do_update(
-            str(attributes['requirement_code']),
-            self._lst_handler_id[0],
-        )
-        self.txtName.do_update(
-            str(attributes['description']),
-            self._lst_handler_id[1],
-        )
-        self.chkDerived.do_update(
-            int(attributes['derived']),
-            self._lst_handler_id[3],
-        )
-        self.cmbRequirementType.do_update(
-            attributes['requirement_type'],
-            self._lst_handler_id[2],
-        )
-        self.txtSpecification.do_update(
-            str(attributes['specification']),
-            self._lst_handler_id[4],
-        )
-        self.txtPageNum.do_update(
-            str(attributes['page_number']),
-            self._lst_handler_id[5],
-        )
-        self.txtFigNum.do_update(
-            str(attributes['figure_number']),
-            self._lst_handler_id[6],
-        )
-        self.cmbPriority.do_update(
-            int(attributes['priority']),
-            self._lst_handler_id[7],
-        )
-        self.cmbOwner.do_update(
-            str(attributes['owner']),
-            self._lst_handler_id[8],
-        )
-        self.chkValidated.do_update(
-            int(attributes['validated']),
-            self._lst_handler_id[9],
-        )
-        if attributes['validated']:
-            self.txtValidatedDate.do_update(
-                str(attributes['validated_date']),
-                self._lst_handler_id[10],
-            )
-        else:
-            self.txtValidatedDate.do_update("", self._lst_handler_id[10])
-
-    def _do_request_update(self, __button):
-        """
-        Send request to save the currently selected Requirement.
-
-        :param __button: the Gtk.ToolButton() that called this method.
-        :type __button: :class:`Gtk.ToolButton`
-        :return: None
-        :rtype: None
-        """
-        self.do_set_cursor(Gdk.CursorType.WATCH)
-        pub.sendMessage(
-            'request_update_requirement',
-            node_id=self._requirement_id,
-        )
-        self.do_set_cursor(Gdk.CursorType.LEFT_PTR)
-
-    def _do_request_update_all(self, __button):
-        """
-        Request to save all the Requirements.
-
-        :param __button: the Gtk.ToolButton() that called this method.
-        :type __button: :class:`Gtk.ToolButton`.
-        :return: None
-        :rtype: None
-        """
-        self.do_set_cursor(Gdk.CursorType.WATCH)
-        pub.sendMessage('request_update_all_requirements')
-        self.do_set_cursor(Gdk.CursorType.LEFT_PTR)
-
-    @staticmethod
-    def _do_select_date(__button, __event, entry):
-        """
-        Request to launch a date selection dialog.
-
-        This method is used to select the validation date for the Requirement.
-
-        :param __button: the ramstk.RAMSTKButton() that called this method.
-        :type __button: :class:`ramstk.gui.gtk.ramstk.RAMSTKButton`
-        :param __event: the Gdk.Event() that called this method.
-        :type __event: :class:`Gdk.Event`
-        :param entry: the Gtk.Entry() that the new date should be displayed in.
-        :type entry: :class:`Gtk.Entry`
-        :return: _date; the date in ISO-8601 (YYYY-mm-dd) format.
-        :rtype: str
-        """
-        _dialog = RAMSTKDateSelect()
-
-        _date = _dialog.do_run()
-        _dialog.do_destroy()
-
-        entry.set_text(str(_date))
-
-        return _date
-
-    def _on_combo_changed(self, combo, index):
-        """
-        Retrieve Gtk.ComboBox() changes and assign to Requirement attribute.
-
-        :param Gtk.CellRendererCombo combo: the Gtk.CellRendererCombo() that
-                                            called this method.
-        :param int index: the position in the Requirement class Gtk.TreeModel()
-                          associated with the data from the calling
-                          Gtk.Entry().
-        :return: None
-        :rtype: None
-        """
-        _dic_keys = {
-            2: 'requirement_type',
-            7: 'priority',
-            8: 'owner',
-        }
-        try:
-            _key = _dic_keys[index]
-        except KeyError:
-            _key = ''
-
-        combo.handler_block(self._lst_handler_id[index])
-
-        _model = combo.get_model()
-        _row = combo.get_active_iter()
-
-        if _key == 'requirement_type':
-            _new_text = [_model.get_value(_row, 0), _model.get_value(_row, 1)]
-        elif _key == 'priority':
-            _new_text = int(_model.get_value(_row, 0))
-        elif _key == 'owner':
-            _new_text = _model.get_value(_row, 0)
-        else:
-            _new_text = ''
-
-        pub.sendMessage(
-            'wvw_editing_requirement',
-            module_id=self._requirement_id,
-            key=_key,
-            value=_new_text,
-        )
-
-        combo.handler_unblock(self._lst_handler_id[index])
-
-    def _on_edit(self, module_id, key, value):  # pylint: disable=unused-argument
-        """
-        Update the Requirement Work View Gtk.Widgets().
-
-        This method updates the Requirement Work View Gtk.Widgets() with
-        changes to the Requirement data model attributes.  This method is
-        called whenever an attribute is edited in a different RAMSTK View.
-
-        :param int module_id: the ID of the Requirement being edited.  This
-                              parameter is required to allow the PyPubSub
-                              signals to call this method and the
-                              request_set_attributes() method in the
-                              RAMSTKDataController.
-        :param str key: the key in the Requirement attributes list of the
-                        attribute that was edited.
-        :param str value: the new text to update the Gtk.Widget() with.
-        :return: None
-        :rtype: None
-        """
-        _dic_switch = {
-            'derived': [self.chkDerived.do_update, 2],
-            'description': [self.txtName.do_update, 3],
-            'figure_number': [self.txtFigNum.do_update, 4],
-            'page_number': [self.txtPageNum.do_update, 6],
-            'specification': [self.txtSpecification.do_update, 10],
-            'validated': [self.chkValidated.do_update, 12],
-            'validated_date': [self.txtValidatedDate.do_update, 13],
-        }
-
-        (_function, _id) = _dic_switch.get(key)
-        _function(value, self._lst_handler_id[_id])
-
-    def _on_focus_out(self, entry, index):
-        """
-        Retrieve RAMSTKEntry() changes and assign to Requirement attributes.
-
-        This method retrieves changes to Requirement attributes through the
-        various Gtk.Widgets() and assign the new data to the appropriate
-        Requirement data model attribute.  This method is called by:
-
-            * Gtk.Entry() 'changed' signal
-            * Gtk.TextView() 'changed' signal
-
-        This method sends the 'wvwEditedRequirement' message.
-
-        :param Gtk.Entry entry: the Gtk.Entry() that called the method.
-        :param int index: the position in the Requirement class Gtk.TreeModel()
-                          associated with the data from the calling
-                          Gtk.Widget().
-        :return: None
-        :rtype: None
-        """
-        _dic_keys = {
-            0: 'requirement_code',
-            1: 'description',
-            4: 'specification',
-            5: 'page_number',
-            6: 'figure_number',
-            10: 'validate_date',
-        }
-        try:
-            _key = _dic_keys[index]
-        except KeyError:
-            _key = ''
-
-        entry.handler_block(self._lst_handler_id[index])
-
-        try:
-            _new_text = str(entry.get_text())
-        except ValueError:
-            _new_text = ''
-
-        pub.sendMessage(
-            'wvw_editing_requirement',
-            module_id=self._requirement_id,
-            key=_key,
-            value=_new_text,
-        )
-
-        entry.handler_unblock(self._lst_handler_id[index])
-
-    def _on_toggled(self, togglebutton, index):
-        """
-        Retrieve Gtk.CheckButton() changes and assign to Requirement attribute.
-
-        :param togglebutton: the Gtk.CheckButton() that called this method.
-        :type togglebutton: :class:`Gtk.CheckButton`
-        :param int index: the position in the Requirement class Gtk.TreeModel()
-                          associated with the data from the calling
-                          Gtk.Entry().
-        :return: None
-        :rtype: None
-        """
-        _dic_keys = {3: 'dervied', 9: 'validate'}
-        try:
-            _key = _dic_keys[index]
-        except KeyError:
-            _key = ''
-
-        togglebutton.handler_block(self._lst_handler_id[index])
-
-        _new_text = int(togglebutton.get_active())
-
-        pub.sendMessage(
-            'wvw_editing_requirement',
-            module_id=self._requirement_id,
-            key=_key,
-            value=_new_text,
-        )
-
-        togglebutton.handler_unblock(self._lst_handler_id[index])
-
-
-class RequirementAnalysis(RAMSTKWorkView):
-    """
-    Display Requirement attribute data in the RAMSTK Work Book.
-
-    The Requirement Analysis Work View displays all the analysis questions and
-    answers for the selected Requirement. The attributes of a Requirement
-    Analysis Work View are:
-
-    :ivar list _lst_clear_a: the list of integers [0, 1] corresponding to the
-                             answers to the Clarity questions.
-    :ivar list _lst_complete_a: the list of integers [0, 1] corresponding to
-                                the answers to the Completeness questions.
-    :ivar list _lst_consistent_a: the list of integers [0, 1] corresponding to
-                                  the answers to the Consistency questions.
-    :ivar list _lst_verifiable_a: the list of integers [0, 1] corresponding to
-                                  the answers to the Verifiability questions.
-    :ivar int _requirement_id: the ID of the Requirement Data Model currently
-                               being controlled.
-    :ivar tvwClear: the :class:`Gtk.RAMSTKTreeView` listing all the Clarity
-                    questions and answers.
-    :ivar tvwComplete: the :class:`Gtk.RAMSTKTreeView` listing all the
-                       Completeness questions and answers.
-    :ivar tvwConsistent: the :class:`Gtk.RAMSTKTreeView` listing all the
-                         Consistency questions and answers.
-    :ivar tvwVerifiable: the :class:`Gtk.RAMSTKTreeView` listing all the
-                         Verifiability questions and answers.
-    """
-
-    _lst_clear = [
-        _(
-            "1. The requirement clearly states what is needed or "
-            "desired.",
-        ),
-        _(
-            "2. The requirement is unambiguous and not open to "
-            "interpretation.",
-        ),
-        _(
-            "3. All terms that can have more than one meaning are "
-            "qualified so that the desired meaning is readily "
-            "apparent.",
-        ),
-        _(
-            "4. Diagrams, drawings, etc. are used to increase "
-            "understanding of the requirement.",
-        ),
-        _(
-            "5. The requirement is free from spelling and "
-            "grammatical errors.",
-        ),
-        _(
-            "6. The requirement is written in non-technical "
-            "language using the vocabulary of the stakeholder.",
-        ),
-        _("7. Stakeholders understand the requirement as written."),
-        _(
-            "8. The requirement is clear enough to be turned over "
-            "to an independent group and still be understood.",
-        ),
-        _(
-            "9. The requirement avoids stating how the problem is "
-            "to be solved or what techniques are to be used.",
-        ),
-    ]
-    _lst_complete = [
-        _(
-            "1. Performance objectives are properly documented "
-            "from the user's point of view.",
-        ),
-        _(
-            "2. No necessary information is missing from the "
-            "requirement.",
-        ),
-        _("3. The requirement has been assigned a priority."),
-        _(
-            "4. The requirement is realistic given the technology "
-            "that will used to implement the system.",
-        ),
-        _(
-            "5. The requirement is feasible to implement given the "
-            "defined project time frame, scope, structure and "
-            "budget.",
-        ),
-        _(
-            "6. If the requirement describes something as a "
-            "'standard' the specific source is cited.",
-        ),
-        _(
-            "7. The requirement is relevant to the problem and its "
-            "solution.",
-        ),
-        _("8. The requirement contains no implied design details."),
-        _(
-            "9. The requirement contains no implied implementation "
-            "constraints.",
-        ),
-        _(
-            "10. The requirement contains no implied project "
-            "management constraints.",
-        ),
-    ]
-    _lst_consistent = [
-        _(
-            "1. The requirement describes a single need or want; "
-            "it could not be broken into several different "
-            "requirements.",
-        ),
-        _(
-            "2. The requirement requires non-standard hardware or "
-            "must use software to implement.",
-        ),
-        _(
-            "3. The requirement can be implemented within known "
-            "constraints.",
-        ),
-        _(
-            "4. The requirement provides an adequate basis for "
-            "design and testing.",
-        ),
-        _(
-            "5. The requirement adequately supports the business "
-            "goal of the project.",
-        ),
-        _(
-            "6. The requirement does not conflict with some "
-            "constraint, policy or regulation.",
-        ),
-        _(
-            "7. The requirement does not conflict with another "
-            "requirement.",
-        ),
-        _(
-            "8. The requirement is not a duplicate of another "
-            "requirement.",
-        ),
-        _("9. The requirement is in scope for the project."),
-    ]
-    _lst_verifiable = [
-        _(
-            "1. The requirement is verifiable by testing, "
-            "demonstration, review, or analysis.",
-        ),
-        _(
-            "2. The requirement lacks 'weasel words' (e.g. "
-            "various, mostly, suitable, integrate, maybe, "
-            "consistent, robust, modular, user-friendly, "
-            "superb, good).",
-        ),
-        _(
-            "3. Any performance criteria are quantified such that "
-            "they are testable.",
-        ),
-        _(
-            "4. Independent testing would be able to determine "
-            "whether the requirement has been satisfied.",
-        ),
-        _(
-            "5. The task(s) that will validate and verify the "
-            "final design satisfies the requirement have been "
-            "identified.",
-        ),
-        _(
-            "6. The identified V&amp;V task(s) have been added to "
-            "the validation plan (e.g., DVP)",
-        ),
-    ]
-
-    def __init__(self, configuration, **kwargs):  # pylint: disable=unused-argument
-        """
-        Initialize the Work View for the Requirement package.
-
-        :param configuration: the RAMSTK Configuration class instance.
-        :type configuration: :class:`ramstk.Configuration.Configuration`
-        """
-        RAMSTKWorkView.__init__(self, configuration, module='Requirement')
-
-        # Initialize private dictionary attributes.
-
-        # Initialize private list attributes.
-        self._lst_clear_a = []
-        self._lst_complete_a = []
-        self._lst_consistent_a = []
-        self._lst_verifiable_a = []
-
-        # Initialize private scalar attributes.
-        self._requirement_id = None
-
-        # Initialize public dictionary attributes.
-
-        # Initialize public list attributes.
-
-        # Initialize public scalar attributes.
-        self.tvwClear = Gtk.TreeView()
-        self.tvwComplete = Gtk.TreeView()
-        self.tvwConsistent = Gtk.TreeView()
-        self.tvwVerifiable = Gtk.TreeView()
-
-        self.__make_ui()
-
-        # Subscribe to PyPubSub messages.
-        pub.subscribe(self._do_clear_page, 'closed_program')
-        pub.subscribe(self._do_load_page, 'selected_requirement')
-
-    def __make_treeview(self, index):
-        """
-        Build the Gtk.TreeView() for the Requirements Analyses.
-
-        :param int index: the index of the Gtk.TreeView() being built.  Used in
-                          callback methods.
-        :return: (_model, _column); the Gtk.ListStore() and
-                                    Gtk.TreeViewColumn() to apply to the RAMSTK
-                                    TreeView().
-        :rtype: tuple
-        """
-        _model = Gtk.ListStore(
-            GObject.TYPE_INT, GObject.TYPE_STRING,
-            GObject.TYPE_INT,
-        )
-
-        _column = Gtk.TreeViewColumn()
-
-        _cell = Gtk.CellRendererText()
-        _cell.set_property('editable', 0)
-        _cell.set_property('visible', 0)
-        _column.pack_start(_cell, True)
-        _column.set_attributes(_cell, text=0)
-
-        _cell = Gtk.CellRendererText()
-        _cell.set_property(
-            'background_rgba', Gdk.RGBA(
-                229.0, 229.0, 229.0,
-                1.0,
-            ),
-        )
-        _cell.set_property('editable', 0)
-        _cell.set_property('wrap-width', 650)
-        _cell.set_property('wrap-mode', Pango.WrapMode.WORD_CHAR)
-        _cell.set_property('yalign', 0.1)
-        _column.pack_start(_cell, True)
-        _column.set_attributes(_cell, markup=1)
-
-        _cell = Gtk.CellRendererToggle()
-        _cell.set_property('activatable', 1)
-        _cell.set_property(
-            'cell-background-rgba',
-            Gdk.RGBA(229.0, 229.0, 229.0, 1.0),
-        )
-        _cell.connect('toggled', self._on_cell_edit, None, 2, _model, index)
-        _column.pack_start(_cell, True)
-        _column.set_attributes(_cell, active=2)
-
-        _column.set_sizing(Gtk.TreeViewColumnSizing.AUTOSIZE)
-
-        return (_model, _column)
-
-    def __make_ui(self):
-        """
-        Build the Requirement Analysis Work View page.
-
-        :return: None
-        :rtype: None
-        """
-        _scrolledwindow = Gtk.ScrolledWindow()
-        _scrolledwindow.set_policy(
-            Gtk.PolicyType.NEVER,
-            Gtk.PolicyType.AUTOMATIC,
-        )
-        _scrolledwindow.add_with_viewport(
-            RAMSTKWorkView._make_buttonbox(
-                self,
-                icons=[],
-                tooltips=[],
-                callbacks=[],
-            ),
-        )
-        self.pack_start(_scrolledwindow, False, False, 0)
-
-        _hpaned = Gtk.HPaned()
-
-        # Create quadrant #1 (upper left) for determining if the
-        # requirement is clear.
-        _vpaned = Gtk.VPaned()
-        _hpaned.pack1(_vpaned, resize=False)
-
-        _vpaned.pack1(self.__make_ui_clear(), resize=False)
-        _vpaned.pack2(self.__make_ui_complete(), resize=False)
-
-        # Create quadrant #2 (upper right) for determining if the
-        # requirement is consistent.
-        _vpaned = Gtk.VPaned()
-        _hpaned.pack2(_vpaned, resize=False)
-
-        _vpaned.pack1(self.__make_ui_consistent(), resize=False)
-        _vpaned.pack2(self.__make_ui_verifiable(), resize=False)
-
-        self.pack_start(_hpaned, True, True, 0)
-
-        _label = Gtk.Label()
-        _label.set_markup("<span weight='bold'>" + _("Analysis") + "</span>")
-        _label.set_alignment(xalign=0.5, yalign=0.5)
-        _label.set_justify(Gtk.Justification.CENTER)
-        _label.set_tooltip_text(_("Analyzes the selected requirement."))
-        _label.show_all()
-        self.hbx_tab_label.pack_start(_label, True, True, 0)
-
-        self.show_all()
-
-    def __make_ui_clear(self):
-        """
-        Build the Gtk.TreeView() to display the questions about clarity.
-
-        :return: _frame; the RAMSTK Frame() containing the clarity questions.
-        :rtype: :class:`ramstk.gui.gtk.ramstk.RAMSTKFrame`
-        """
-        _scrollwindow = Gtk.ScrolledWindow()
-        _scrollwindow.set_policy(
-            Gtk.PolicyType.AUTOMATIC,
-            Gtk.PolicyType.AUTOMATIC,
-        )
-        _scrollwindow.add(self.tvwClear)
-
-        _frame = RAMSTKFrame(label=_("Clarity of Requirement"))
-        _frame.set_shadow_type(Gtk.ShadowType.ETCHED_OUT)
-        _frame.add(_scrollwindow)
-
-        (_model, _column) = self.__make_treeview(0)
-        self.tvwClear.set_model(_model)
-        self.tvwClear.set_headers_visible(False)
-        self.tvwClear.append_column(_column)
-
-        _model.clear()
-        for _index, _answer in enumerate(self._lst_clear):
-            _model.append(
-                [_index, "<span weight='bold'>" + _answer + "</span>", 0],
-            )
-
-        return _frame
-
-    def __make_ui_complete(self):
-        """
-        Build the Gtk.TreeView() to display questions about completeness.
-
-        :return: _frame; the RAMSTK Frame() containing the completeness
-                         questions.
-        :rtype: :class:`ramstk.gui.gtk.ramstk.RAMSTKFrame`
-        """
-        _scrollwindow = Gtk.ScrolledWindow()
-        _scrollwindow.set_policy(
-            Gtk.PolicyType.AUTOMATIC,
-            Gtk.PolicyType.AUTOMATIC,
-        )
-        _scrollwindow.add(self.tvwComplete)
-
-        _frame = RAMSTKFrame(label=_("Completeness of Requirement"))
-        _frame.set_shadow_type(Gtk.ShadowType.ETCHED_OUT)
-        _frame.add(_scrollwindow)
-
-        (_model, _column) = self.__make_treeview(1)
-        self.tvwComplete.set_model(_model)
-        self.tvwComplete.set_headers_visible(False)
-        self.tvwComplete.append_column(_column)
-
-        _model.clear()
-        for _index, _answer in enumerate(self._lst_complete):
-            _model.append(
-                [_index, "<span weight='bold'>" + _answer + "</span>", 0],
-            )
-
-        return _frame
-
-    def __make_ui_consistent(self):
-        """
-        Build the Gtk.TreeView() to display the questions about consistency.
-
-        :return: _frame; the RAMSTK Frame() containing the consistency
-                         questions.
-        :rtype: :class:`ramstk.gui.gtk.ramstk.RAMSTKFrame`
-        """
-        _scrollwindow = Gtk.ScrolledWindow()
-        _scrollwindow.set_policy(
-            Gtk.PolicyType.AUTOMATIC,
-            Gtk.PolicyType.AUTOMATIC,
-        )
-        _scrollwindow.add(self.tvwConsistent)
-
-        _frame = RAMSTKFrame(label=_("Consistency of Requirement"))
-        _frame.set_shadow_type(Gtk.ShadowType.ETCHED_OUT)
-        _frame.add(_scrollwindow)
-
-        (_model, _column) = self.__make_treeview(2)
-        self.tvwConsistent.set_model(_model)
-        self.tvwConsistent.set_headers_visible(False)
-        self.tvwConsistent.append_column(_column)
-
-        _model.clear()
-        for _index, _answer in enumerate(self._lst_consistent):
-            _model.append(
-                [_index, "<span weight='bold'>" + _answer + "</span>", 0],
-            )
-
-        return _frame
-
-    def __make_ui_verifiable(self):
-        """
-        Build the Gtk.TreeView() to display questions about verifiability.
-
-        :return: _frame; the RAMSTK Frame() containing the verifiability
-                         questions.
-        :rtype: :class:`ramstk.gui.gtk.ramstk.RAMSTKFrame`
-        """
-        _scrollwindow = Gtk.ScrolledWindow()
-        _scrollwindow.set_policy(
-            Gtk.PolicyType.AUTOMATIC,
-            Gtk.PolicyType.AUTOMATIC,
-        )
-        _scrollwindow.add(self.tvwVerifiable)
-
-        _frame = RAMSTKFrame(label=_("Verifiability of Requirement"))
-        _frame.set_shadow_type(Gtk.ShadowType.ETCHED_OUT)
-        _frame.add(_scrollwindow)
-
-        (_model, _column) = self.__make_treeview(3)
-        self.tvwVerifiable.set_model(_model)
-        self.tvwVerifiable.set_headers_visible(False)
-        self.tvwVerifiable.append_column(_column)
-
-        _model.clear()
-        for _index, _answer in enumerate(self._lst_verifiable):
-            _model.append(
-                [_index, "<span weight='bold'>" + _answer + "</span>", 0],
-            )
-
-        return _frame
-
-    def _do_clear_page(self):
-        """
-        Clear the contents of the page.
-
-        :return: None
-        :rtype: None
-        """
-        for _treeview in [
-                self.tvwClear, self.tvwComplete, self.tvwConsistent,
-                self.tvwVerifiable,
-        ]:
-            _model = _treeview.get_model()
-            _columns = _treeview.get_columns()
-            for _column in _columns:
-                _treeview.remove_column(_column)
-
-            _model.clear()
-
-    def _do_load_page(self, attributes):
-        """
-        Load the Requirements analysis page.
-
-        :return: None
-        :rtype: None
-        """
-        self._revision_id = attributes['revision_id']
-        self._requirement_id = attributes['requirement_id']
-
-        RAMSTKWorkView.on_select(
-            self,
-            title=_("Analyzing Requirement {0:s} - {1:s}").format(
-                str(attributes['requirement_code']),
-                str(attributes['description']),
-            ),
-        )
-
-        # Load the Requirement analyses answers.  It's easiest to pack the
-        # answers into a list and iterate for each tree.
-        self._lst_clear_a = [
-            attributes['q_clarity_0'], attributes['q_clarity_1'],
-            attributes['q_clarity_2'], attributes['q_clarity_3'],
-            attributes['q_clarity_4'], attributes['q_clarity_5'],
-            attributes['q_clarity_6'], attributes['q_clarity_7'],
-            attributes['q_clarity_8'],
-        ]
-        _model = self.tvwClear.get_model()
-        _row = _model.get_iter_first()
-        for _answer in self._lst_clear_a:
-            _model.set_value(_row, 2, _answer)
-            _row = _model.iter_next(_row)
-
-        self._lst_complete_a = [
-            attributes['q_complete_0'], attributes['q_complete_1'],
-            attributes['q_complete_2'], attributes['q_complete_3'],
-            attributes['q_complete_4'], attributes['q_complete_5'],
-            attributes['q_complete_6'], attributes['q_complete_7'],
-            attributes['q_complete_8'], attributes['q_complete_9'],
-        ]
-        _model = self.tvwComplete.get_model()
-        _row = _model.get_iter_first()
-        for _answer in self._lst_complete_a:
-            _model.set_value(_row, 2, _answer)
-            _row = _model.iter_next(_row)
-
-        self._lst_consistent_a = [
-            attributes['q_consistent_0'], attributes['q_consistent_1'],
-            attributes['q_consistent_2'], attributes['q_consistent_3'],
-            attributes['q_consistent_4'], attributes['q_consistent_5'],
-            attributes['q_consistent_6'], attributes['q_consistent_7'],
-            attributes['q_consistent_8'],
-        ]
-        _model = self.tvwConsistent.get_model()
-        _row = _model.get_iter_first()
-        for _answer in self._lst_consistent_a:
-            _model.set_value(_row, 2, _answer)
-            _row = _model.iter_next(_row)
-
-        self._lst_verifiable_a = [
-            attributes['q_verifiable_0'], attributes['q_verifiable_1'],
-            attributes['q_verifiable_2'], attributes['q_verifiable_3'],
-            attributes['q_verifiable_4'], attributes['q_verifiable_5'],
-        ]
-        _model = self.tvwVerifiable.get_model()
-        _row = _model.get_iter_first()
-        for _answer in self._lst_verifiable_a:
-            _model.set_value(_row, 2, _answer)
-            _row = _model.iter_next(_row)
-
-    def _do_request_update(self, __button):
-        """
-        Send request to save the currently selected Requirement.
-
-        :param __button: the Gtk.ToolButton() that called this method.
-        :type __button: :class:`Gtk.ToolButton`
-        :return: None
-        :rtype: None
-        """
-        self.do_set_cursor(Gdk.CursorType.WATCH)
-        pub.sendMessage(
-            'request_update_requirement',
-            node_id=self._requirement_id,
-        )
-        self.do_set_cursor(Gdk.CursorType.LEFT_PTR)
-
-    def _do_request_update_all(self, __button):
-        """
-        Request to save all the Requirements.
-
-        :param __button: the Gtk.ToolButton() that called this method.
-        :type __button: :class:`Gtk.ToolButton`.
-        :return: None
-        :rtype: None
-        """
-        self.do_set_cursor(Gdk.CursorType.WATCH)
-        pub.sendMessage('request_update_all_requirements')
-        self.do_set_cursor(Gdk.CursorType.LEFT_PTR)
-
-    def _on_cell_edit(self, cell, path, new_text, position, model, index):
-        """
-        Handle edits of the Requirement Analysis RAMSTKTreeview().
-
-        :param cell: the Gtk.CellRendererToggle() that was toggled.
-        :type cell: :class:`Gtk.CellRendererToggle`
-        :param str path: the Gtk.TreeView() path of the Gtk.CellRenderer()
-                         that was edited.
-        :param model: the Gtk.TreeModel() for the Gtk.Treeview() that is being
-                      edited.
-        :type model: :class:`Gtk.TreeModel`
-        :param int index: the index of the Requirement analysis Gtk.Treeview()
-                          questions being answered.  Indices are:
-
-                             * 0 = clarity
-                             * 1 = completeness
-                             * 2 = consistency
-                             * 3 = verifiability
-
-        :return: None
-        :rtype: None
-        """
-        position = model[path][0]
-
-        new_text = boolean_to_integer(not cell.get_active())
-        model[path][2] = new_text
-
-        try:
-            if index == 0:
-                self._lst_clear_a[position] = new_text
-                _key = 'q_clarity_{0:d}'.format(position)
-            elif index == 1:
-                self._lst_complete_a[position] = new_text
-                _key = 'q_complete_{0:d}'.format(position)
-            elif index == 2:
-                self._lst_consistent_a[position] = new_text
-                _key = 'q_consistent_{0:d}'.format(position)
-            elif index == 3:
-                self._lst_verifiable_a[position] = new_text
-                _key = 'q_verifiable_{0:d}'.format(position)
-        except IndexError:
-            _debug_msg = (
-                "RAMSTK ERROR: No position {0:d} in Requirements "
-                "Analysis question list for index {1:d}."
-            ).format(
-                position, index,
-            )
-            self.RAMSTK_CONFIGURATION.RAMSTK_DEBUG_LOG.error(_debug_msg)
-
-        pub.sendMessage(
-            'wvw_editing_requirement',
-            module_id=self._requirement_id,
-            key=_key,
-            value=new_text,
-        )
